# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  name: string;
  email: string;
  phone: string;
  course: string;
  status: string;
  baseAmount: number;
  gstAmount: number;
  totalAmount: number;
  /** @nullable */
  utrNumber?: string | null;
  createdAt: string;
}