# lib/api-zod/src/generated/types/enrollmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnrollmentInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 2 */
  course: string;
}