# lib/api-zod/src/generated/types/paymentConfirmInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentConfirmInput {
  /** @minLength 4 */
  utrNumber: string;
}